Move error handlers to separate file
